avoid out-of-bounds panic in rbf
We've seen this happen with relatively large datasets with a relatively low
max-file-size. The solution we came up with was to increase the max-file-size
config option, which works, but we still don't want there to be a panic if we
hit this again.

see https://molecula.atlassian.net/browse/FB-1381 for more information.