delete things from holder after removal from etcd
This way we can't get into an inconsistent state.

Consider this example:
You have a 3-node cluster, nodes A, B and C.
You create an index "blah" while all three nodes are up.
Nodes B and C go down.
You attempt to delete the index. It is removed from node A's holder, but is not removed from nodes B and C.
When nodes B and C are restarted, the schema still shows this "blah" index.
If you attempt to delete the index from node A, you receive an index not found error, but the schema indicates the index exists.

With this change however, when you first attempt to delete the index, it is not removed from the holder until there is enough nodes up to achieve consensus.

The same situation applies to fields and views.