Distinguish between three copies of `fakeidp` service (#2192)
There are currently three copies of a package called `fakeidp` in the
featurebase repo:

- ./idk/fakeidp/go.mod
- ./internal/clustertests/fakeidp/go.mod
- ./qa/fakeidp/go.mod

All three have a `go.mod` file. While this is supported under golang's
new Workspace support, what's not supported is that the modules share
the same name (in this case "fakeidp"). This commit is a sort of
temporary fix which renames the module for two of the instances. This
prevents, for example, VSCode with workspace support enabled, from
barfing.

By the way, one can enable VSCode workspace support with the following
setting:

```
// gopls
"gopls": {
    "build.experimentalWorkspaceModule": true
}
```

Also...

This commit fixes the `make testv` target. It's probably not used
anywhere (which I'm assuming because it was broken), but it's a handle
target, so now it will list and run tests against all packages found in
the repo, including the root package.