add "only" clauses to IDK tests
The IDK tests should be run only when there's changes in the IDK
or client directories.

Also the shard transactional test should be run ever.

The "optional: true" flag is a fascinating quirk allowing you to
express that, *if* a job exists, we should wait for it, but if the
job doesn't exist, that's also acceptable.